+++ api/v1alpha1/frequi_types.go
@@ -132,6 +132,7 @@ type FreqUIStatus struct {
 }
 
 // +kubebuilder:object:root=true
+// +kubebuilder:deprecatedversion:warning="freqtrade.io/v1alpha1 FreqUI is deprecated; use freqtrade.io/v1beta1. See https://github.com/ark-sys/freqtrade-operator#upgrading-to-v1beta1"
 // +kubebuilder:subresource:status
 // +kubebuilder:printcolumn:name="Host",type=string,JSONPath=`.spec.host`
 // +kubebuilder:printcolumn:name="Phase",type=string,JSONPath=`.status.phase`